requirements.txt
flask
python-dotenv
google-genai
zai-sdk
gunicorn
flask-cors